ui/frontend: add ProfileProvider and ProfileConsumer
As with our other frontends, we have a "ProfileProvider" component which
fetches the current user's profile and a "ProfileConsumer" component
which other components can use to obtain the current user profile.

Add ProfileProvider to the index page so that all rendered components
have the option of retrieving the current profile.