ui/frontend: add UpdatePreference component
Add a component which can be used to update the current preference for a
user. It provides a "draft" preference which can be modified via a
function passed to its children. When this preference is ready to be
submitted, children may update the preference via another function
passed to them.